toml reading solved

+++ src/Plugin/Components/TransformComponent.h
@@ -6,11 +6,11 @@
 struct PluginTransformComponent {
   glm::vec2 position;
   glm::vec2 scale;
-  double rotation;
+  float rotation;
 
   PluginTransformComponent(glm::vec2 position = glm::vec2(0, 0),
                            glm::vec2 scale = glm::vec2(1.0, 1.0),
-                           double rotation = 0.0f)
+                           float rotation = 0.0f)
       : position(position), scale(scale), rotation(rotation) {}
 };
